requirements.txt
streamlit
transformers
torch
soundfile
torchaudio
googletrans==4.0.0-rc1
Pillow
timm